+++ src/main/java/pro/zackpollard/telegrambot/api/internal/event/ListenerRegistryImpl.java
@@ -67,6 +67,18 @@ public class ListenerRegistryImpl implements ListenerRegistry {
         }
     }
 
+    public void unregister(Listener listener) {
+
+        for (Method m : listener.getClass().getDeclaredMethods()) {
+
+            Class<?>[] classes = m.getParameterTypes();
+            if(classes.length == 1) {
+
+                listenerByContent.computeIfPresent(classes[0], (c, s) -> (s.remove(listener) && s.size() == 0 ? null : s));
+            }
+        }
+    }
+
     public void callEvent(Event event) {
         BiConsumer<Listener, Event> invoker = (BiConsumer<Listener, Event>) this.invokers.get(event.getClass());
         listenerByContent.getOrDefault(event.getClass(), Collections.emptySet()).forEach(listener -> invoker.accept(listener, event));